backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
PyMuPDF==1.23.8
pikepdf==8.11.2
python-docx==1.1.0
pdf2image==1.17.0
Pillow==10.2.0
pytesseract==0.3.10
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-magic-bin==0.4.14
